improvements to tests
